+++ functions.py
@@ -6,6 +6,12 @@ def minus(a, b):
     """Subtract b from a."""
     return a - b
 
+def modulo(a, b):
+    """a mod b"""
+    if b == 0:
+        print("Can't operate!!")
+    else:
+        return a % b
 def multiply(a, b):
     """Multiply two numbers"""
     return a * b